+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.3.8] - 2019-03-13
+### Added
+- `gpsLocValid` parameter in `UpdateContainer`
+
+### Changed
+- few more functions moved from `ActionTools` to Kotlin based `ActionBasics`
+
 ## [0.3.7] - 2019-02-28
 ### Changed
 - united adding attachments to `GeoDataExtra`